+++ assets/modules/cellType.mjs
@@ -1,0 +1,9 @@
+export const CellType = Object.freeze({
+    PLAINS: Symbol("plains"),
+    FOREST: Symbol("forest"),
+    MOUNTAIN: Symbol("mountain"),
+    SNOWY_MOUNTAIN: Symbol("snowy_mountain"),
+    SHALLOW_WATER: Symbol("shallow_water"),
+    DEEP_WATER: Symbol("deep_water"),
+    DESERT: Symbol("desert")
+});