+++ assets/modules/worldGenerators/generator.mjs
@@ -1,0 +1,12 @@
+import { World } from "../world.mjs";
+
+export class Generator {
+    /**
+     * Generate a new world
+     * @param {object} options Options for generation
+     * @returns {World} A whole new world
+     */
+    static create({}) {
+        throw new Error("Method not implemented.");
+    }
+}