+++ CHANGELOG.md
@@ -6,3 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
 ## [Unreleased]
+
+### Added
+
+- Cells
+- Cell types
+- Worlds
+- World generators
+- Perlin-based world generator with octaves
+- Perlin-based world generator with preset fading layers
+- Canvas drawer